+++ tests/CoreParserTest.php
@@ -11,6 +11,7 @@ use Serhii\GoodbyeHtml\Ast\HtmlStatement;
 use Serhii\GoodbyeHtml\Ast\IfExpression;
 use Serhii\GoodbyeHtml\Ast\IntegerLiteral;
 use Serhii\GoodbyeHtml\Ast\LoopExpression;
+use Serhii\GoodbyeHtml\Ast\NullLiteral;
 use Serhii\GoodbyeHtml\Ast\PrefixExpression;
 use Serhii\GoodbyeHtml\Ast\StringLiteral;
 use Serhii\GoodbyeHtml\Ast\TernaryExpression;
@@ -56,14 +57,6 @@ class CoreParserTest extends TestCase
         $this->assertSame('<div class="nice"></div>', $stmt->string());
     }
 
-    public static function providerForTestPrefixExpressions(): array
-    {
-        return [
-            ['{{ -4 }}', '-', 4],
-            ['{{ -284 }}', '-', 284],
-        ];
-    }
-
     /**
      * @dataProvider providerForTestBooleanExpressions
      */
@@ -81,6 +74,14 @@ class CoreParserTest extends TestCase
 
         $this->assertInstanceOf(BooleanExpression::class, $prefix);
         $this->assertSame($expect, $prefix->string());
+    }
+
+    public static function providerForTestBooleanExpressions(): array
+    {
+        return [
+            ['{{ true }}', 'true'],
+            ['{{ false }}', 'false'],
+        ];
     }
 
     public function testParsingIfExpression(): void
@@ -284,15 +285,16 @@ class CoreParserTest extends TestCase
         $prefix = $program->statements[0]->expression;
 
         $this->assertInstanceOf(PrefixExpression::class, $prefix);
-        $this->testInteger($prefix->right, $expect);
         $this->assertSame($operator, $prefix->operator);
     }
 
-    public static function providerForTestBooleanExpressions(): array
+    public static function providerForTestPrefixExpressions(): array
     {
         return [
-            ['{{ true }}', 'true'],
-            ['{{ false }}', 'false'],
+            ['{{ -4 }}', '-', 4],
+            ['{{ -284 }}', '-', 284],
+            ['{{ !true }}', '!', false],
+            ['{{ !false }}', '!', true],
         ];
     }
 
@@ -332,5 +334,23 @@ class CoreParserTest extends TestCase
     {
         self::assertInstanceOf(FloatLiteral::class, $float);
         self::assertSame($val, $float->value, "Float must have value '{$val}', got: '{$float->value}'");
+    }
+
+    public function testParsingNull(): void
+    {
+        $input = '{{ null }}';
+
+        $lexer = new Lexer($input);
+        $parser = new CoreParser($lexer);
+
+        $program = $parser->parseProgram();
+
+        $this->checkForErrors($parser, $program->statements, 1);
+
+        /** @var NullLiteral $null */
+        $null = $program->statements[0]->expression;
+
+        self::assertInstanceOf(NullLiteral::class, $null);
+        self::assertSame('null', $null->token->literal);
     }
 }